system-image-upgrader: implement support for LVM migration
If enabled by setting ro.systemimage.use_lvm=true property, it will attempt
to slightly shrink userdata and convert it in-place to LVM logical
volume.

Then the system partition is going to be migrated to LVM
as well and expanded to accommodate for a larger rootfs on older devices.

(H16 JAMI: fold in 23557e95 + do previous ubports/Android.mk awk symlink
in init.rc for now at least)

Change-Id: I2006b328a50d6808254102ba1083b9d96efe1151